+++ Assets/UniGLTF/Runtime/MeshUtility/BlendShapeReport.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ab7e7ef8eaac5b4593a6a1102c15012
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
