+++ Assets/UniGLTF/Runtime/MeshUtility/MeshFreezer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09547456fd06d7b419d7553beba6f58a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
